New nngnorb mail backend
lisp/nngnorb.el: Define a "nngnorb" mail backend, mostly copy and pasted
		 from nnnil.

lisp/gnorb.el: Require the above file.

lisp/gnorb-org.el: New function `gnorb-org-view'. Call this on a subtree
		 to open an ephemeral gnus group containing all the
		 messages linked to from within the subtree's text. This
		 calls directly to `gnorb-gnus-search-messages'.

lisp/gnorb-gnus.el: New function `gnorb-gnus-search-messages', initiates
		 the actual search. Analogous to
		 `gnus-group-make-nnir-group'. Works in the *Server*
		 buffer, on the nngnorb server line, but can't be used
		 from the *Group* buffer, as the nngnorb backend has no
		 groups.

Once the search is initiated, the ephemeral group construction process
eventually calls `nnir-run-gnorb', with the search query string. This
string can be one of two things. The first is an Org id value, which is
what happens when you start out using `gnorb-org-view'. The heading with
that value is scanned for links to gnus messages. The second is an
agenda search string, with tags/todo matching and all that. In that case
all the headings which match the search are scanned for message links.
It might be nice to make those groups persistent at some point.