Register sent messages so we can view them later
When using gnorb-org-view on a subtree, it previously only scanned for
gnus links to show in the nnir buffer. But we'd really like to see our
sent messages as well, ie the ones which are recorded in the
GNORB_MSG_IDS property. The registry is currently the most reliable way
of doing that, so we go to some lengths to make sure sent messages are
stored in the registry, and that we can get them out again when we use
gnorb-org-view.

The current major failure is that, if we don't use a local archive
method, and our sent message come back to us via some unknown imap
group (a la gmail's [Gmail]Sent Mail), we can't find them again.

* lisp/gnorb-gnus.el (gnorb-gnus-check-outgoing-headers): When we have a
  real gcc, insert the message into the registry. Additionally, give
  entries a 'gnorb-ids key, and add that key to the
  gnus-registry-extra-entries-precious list, preventing them from
  getting pruned.
* lisp/gnorb-utils.el (gnorb-msg-id-to-link): New function that tries to
  use a message-id to find the full location of a message. Currently
  only uses the registry, but other methods could be added.
* lisp/nngnorb.el (nnir-run-gnorb): When scanning subtrees for messages
  to display, also try to pick up messages referred to in GNORB_MSG_ID
  properties, using the previous new function to find their real
  location.