Begin shift to using the registry
All the property-based stuff is still in place, as is the association
hashtable. But as a first step we're now using the registry to track
incoming and outgoing messages that are related to Org TODOs. Once this
is working and debugged, then start refactoring other parts of the code
to use this and not the hashtable.

* lisp/gnorb-gnus.el (gnorb-gnus-capture-registry): New function. When
  capturing Org headings from Gnus messages, store the Org id in the
  message's registry entry.
  (gnorb-gnus-make-registry-entry): New function. Make a registry entry
  from message headers.
  (gnorb-gnus-outgoing-make-todo-1): Manually make registry entries from
  new outgoing message.
  (gnorb-gnus-incoming-do-todo): Make registry entries from incoming
  messages.

* lisp/gnorb-utils.el (gnorb-trigger-todo-action): After sending
  relevant messages, add the Org id of triggered headlines to that
  message's registry entry.